start doing more can abstraction haha
