+++ boards/stm32f767zi/src/tasks/heartbeat_coordinator.rs
@@ -1,6 +1,6 @@
 use crate::emergency;
 
-use super::can_sender::CAN_SEND;
+use super::can::CAN_SEND;
 use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, channel::Channel};
 use embassy_time::{Duration, Instant, Timer};
 use hyped_core::{
