+++ boards/stm32f767zi/src/tasks/heartbeats_responder.rs
@@ -1,4 +1,4 @@
-use super::can_sender::CAN_SEND;
+use super::can::CAN_SEND;
 use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, channel::Channel};
 use embassy_time::{Duration, Timer};
 use hyped_core::comms::{boards::Board, heartbeat::Heartbeat, messages::CanMessage};
